+++ src/Environment.fs
@@ -4,10 +4,12 @@ open System
 open System.Text.RegularExpressions
 
 // inspired by http://www.codesuji.com/2018/02/28/F-and-DotEnv/
-let private loadDotEnv error envFile =
-    if IO.File.Exists(envFile)
-    then
-        IO.File.ReadAllLines(envFile)
+let loadFromFile filePath =
+    result {
+        if IO.File.Exists(filePath) |> not then
+            return! Error (sprintf "File %s does not exists." filePath)
+
+        IO.File.ReadAllLines(filePath)
         |> Seq.iter (fun line ->
             let lineWithoutComments = Regex.Replace(line, "#.*", "")
             let parts =
@@ -19,25 +21,14 @@ let private loadDotEnv error envFile =
                 Environment.SetEnvironmentVariable(key, value.Trim([|'"'; '''|]))
             | _ -> ()
         )
-    else
-        error (sprintf "File %s does not exists." envFile)
+    }
 
-let getEnvs error envFile =
-    loadDotEnv error envFile
-
+let getEnvs () =
     // https://stackoverflow.com/questions/12014179/f-get-environment-variables-as-a-generic-collection
     System.Environment.GetEnvironmentVariables()
     |> Seq.cast<System.Collections.DictionaryEntry>
     |> Seq.map (fun d -> d.Key :?> string, d.Value :?> string)
     |> Map
-
-let tryGetEnv (envs: Map<string, string>) key =
-    envs |> Map.tryFind key
-
-let getEnv (envs: Map<string, string>) key =
-    match key |> tryGetEnv envs with
-    | Some value -> value
-    | _ -> failwithf "[Error]: ENV variable \"%s\" was not defined." key
 
 let private combine onSame (currentEnvs: Map<string, string>) (newEnvs: Map<string, string>): Map<string, string> =
     newEnvs